basic structure for index.cpp and move methods from readCollection.cpp
